+++ examples/weather/cloesce.config.ts
@@ -5,6 +5,7 @@ const config = defineConfig({
     srcPaths: ["./src/data"],
     workersUrl: "http://localhost:5000/api",
     migrationsPath: "./customDir",
+    wranglerConfigFormat: "jsonc",
 });
 
 config.model(Weather, builder => {